Rename openspecrtra to openspectra
